Endret person_id fra int til streng i Avregningsretur

+++ src/main/kotlin/no/nav/sokos/spk/mottak/domain/converter/AvregningConverter.kt
@@ -36,7 +36,7 @@ object AvregningConverter {
             datoValutering = datoValutert.toLocalDateStringOrEmpty(),
             konto = konto,
             motId = delytelseId,
-            personId = fagSystemId.toInt(),
+            personId = fagSystemId,
             kreditorRef = kreditorRef,
             datoOpprettet = LocalDateTime.now(),
             opprettetAv = systemId,
